work around 5.10.0 not saving source of eval with sub
Under the debugger, the source of an eval is supposed to be saved if a
sub was defined in it.  perl 5.10.0 doesn't appear to do this, and I
don't see any possible workaround.  This doesn't seem to be a problem in
either 5.8 or 5.10.1.  Mark the test for this as TODO.