restore threads::tid if required
We can be triggered inside a Safe compartment, where most of the stash
entries are gone.  Carp may have early bound the threads::tid glob, so
it detects the sub existing, but when calling it as a method the stash
is gone and errors.

Detect if the tid sub is missing and temporarily restore it from the original
glob.