extract original can and isa subrefs using B
If the UNIVERSAL::isa or UNIVERSAL::can modules are loaded, the original
methods are wrapped and will cause us problems.  There are variables we
can set so they stay our of our way, but it would be preferrable to get
access to the original subs.

Using B.pm, we can access the pads of the wrapper subs and find the
captured original.  For now, maintain the variable setting version as a
fallback.